refactor(client): Extract `Navbar` & `Search` into components
